+++ src/tests/automated_tests/unit/test_unit_status.py
@@ -9,6 +9,10 @@ import requests
 import os
 from markers import skipif_not_uses_gateway
 
+### SSL parameters
+SSL_ENABLED = (os.environ.get("F7T_SSL_ENABLED","false").lower() == "true")
+SSL_CRT = os.environ.get("F7T_SSL_CRT", "")
+SSL_PATH = "../../../deploy/test-build"
 
 FIRECREST_URL = os.environ.get("FIRECREST_URL")
 USE_GATEWAY  = (os.environ.get("USE_GATEWAY","false").lower() == "true")
@@ -16,14 +20,14 @@ USE_GATEWAY  = (os.environ.get("USE_GATEWAY","false").lower() == "true")
 if FIRECREST_URL and USE_GATEWAY:
 	STATUS_URL = os.environ.get("FIRECREST_URL") + "/status"
 else:
-    STATUS_URL = os.environ.get("F7T_STATUS_URL")
+	F7T_SCHEME_PROTOCOL = ("https" if SSL_ENABLED else "http")
+	STATUS_HOST = os.environ.get("F7T_STATUS_HOST","127.0.0.1") 
+	STATUS_PORT = os.environ.get("F7T_STATUS_PORT","5001")
+	STATUS_URL = f"{F7T_SCHEME_PROTOCOL}://{STATUS_HOST}:{STATUS_PORT}"
 
-SYSTEMS = os.environ.get("F7T_SYSTEMS_PUBLIC").strip('\'"').split(";")
+SYSTEMS = os.environ.get("F7T_SYSTEMS_PUBLIC_NAME").strip('\'"').split(";")
 
-### SSL parameters
-USE_SSL = os.environ.get("F7T_USE_SSL", False)
-SSL_CRT = os.environ.get("F7T_SSL_CRT", "")
-SSL_PATH = "../../../deploy/test-build"
+print(f"STATUS_URL: {STATUS_URL}")
 
 @skipif_not_uses_gateway
 @pytest.mark.parametrize("system",SYSTEMS)
@@ -53,7 +57,7 @@ STATUS_CODES_SERVICES.append(("not-a-service",404))
 @pytest.mark.parametrize("system,status_code", STATUS_CODES_SYSTEMS)
 def test_status_system(system, status_code, headers):
 	url = f"{STATUS_URL}/systems/{system}"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(resp.content)
 	# assert 'description' in resp.json()
 	assert status_code == resp.status_code
@@ -61,7 +65,7 @@ def test_status_system(system, status_code, headers):
 @skipif_not_uses_gateway
 def test_status_systems(headers):
 	url = f"{STATUS_URL}/systems"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(resp.content)
 	assert 'description' in resp.json()
 
@@ -69,7 +73,7 @@ def test_status_systems(headers):
 @pytest.mark.parametrize("system,fs_name,status_code", STATUS_CODES_FS)
 def test_status_filesystems(system,fs_name,status_code,headers):
 	url = f"{STATUS_URL}/filesystems/{system}"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	# /home is OK, and /scratch is wrong
 	print(resp.content)
 
@@ -86,7 +90,7 @@ def test_status_filesystems(system,fs_name,status_code,headers):
 @pytest.mark.parametrize("system,fs_name,status_code", STATUS_CODES_FS)
 def test_status_all_filesystems(system,fs_name,status_code,headers):
 	url = f"{STATUS_URL}/filesystems"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	# /home is OK, and /scratch is wrong
 	print(resp.content)
 	assert "out" in resp.json()
@@ -106,7 +110,7 @@ def test_status_all_filesystems(system,fs_name,status_code,headers):
 @pytest.mark.parametrize("service,status_code", STATUS_CODES_SERVICES)
 def test_status_service(service, status_code, headers):
 	url = f"{STATUS_URL}/services/{service}"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(resp.json())
 	# assert 'description' in resp.json()
 	assert status_code == resp.json()["status_code"]
@@ -114,7 +118,7 @@ def test_status_service(service, status_code, headers):
 @skipif_not_uses_gateway
 def test_status_services(headers):
 	url = f"{STATUS_URL}/services"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(resp.content)
 	# print(resp.json())
 	assert 'description' in resp.json()
@@ -123,7 +127,7 @@ def test_status_services(headers):
 def test_parameters(headers):
 	print(STATUS_URL)
 	url = f"{STATUS_URL}/parameters"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(resp.content)
 	assert resp.status_code == 200