+++ src/tests/automated_tests/unit/test_unit_tasks.py
@@ -11,18 +11,24 @@ import os
 # from markers import host_environment_test
 from markers import skipif_uses_gateway, skipif_not_uses_gateway
 
+### SSL parameters
+SSL_ENABLED = (os.environ.get("F7T_SSL_ENABLED","false").lower() == "true")
+SSL_CRT = os.environ.get("F7T_SSL_CRT", "")
+SSL_PATH = "../../../deploy/test-build"
+
 FIRECREST_URL = os.environ.get("FIRECREST_URL","")
 USE_GATEWAY  = (os.environ.get("USE_GATEWAY","false").lower() == "true")
 
 if FIRECREST_URL and USE_GATEWAY:
 	TASKS_URL = os.environ.get("FIRECREST_URL") + "/tasks"
 else:
-    TASKS_URL = os.environ.get("F7T_TASKS_URL")
+    F7T_SCHEME_PROTOCOL = ("https" if SSL_ENABLED else "http")
+    
+    TASKS_HOST = os.environ.get("F7T_TASKS_HOST","127.0.0.1") 
+    TASKS_PORT = os.environ.get("F7T_TASKS_PORT","5003")
+    TASKS_URL = f"{F7T_SCHEME_PROTOCOL}://{TASKS_HOST}:{TASKS_PORT}"
 
-### SSL parameters
-USE_SSL = os.environ.get("F7T_USE_SSL", False)
-SSL_CRT = os.environ.get("F7T_SSL_CRT", "")
-SSL_PATH = "../../../deploy/test-build"
+print(f"TASKS_URL: {TASKS_URL}")
 
 INVALID_CODE1 = "9999"
 INVALID_CODE2 = "47777"
@@ -61,7 +67,7 @@ STATUS_CODES = [(QUEUED, "queued", 200), (PROGRESS, "progress", 200), (SUCCESS, 
 # helper function to create a task
 def create_task(headers):
 	url = "{}".format(TASKS_URL)
-	resp = requests.post(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False) )
+	resp = requests.post(url, headers=headers, verify=False)
 	print(resp.content)
 	print(url)
 	return resp
@@ -77,7 +83,7 @@ def test_list_tasks(task_list, status_code, headers):
 		task_list_query = f"?tasks={task_list}"
 	
 	url = f"{TASKS_URL}/{task_list_query}"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(json.dumps(resp.json(),indent=2))
 	print(url)
 	assert resp.status_code == status_code
@@ -96,7 +102,7 @@ def test_get_task(headers):
 	resp = create_task(headers)
 	hash_id = resp.json()["hash_id"]
 	url = "{}/{}".format(TASKS_URL, hash_id)
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(json.dumps(resp.json(),indent=2))
 	assert resp.status_code == 200
 
@@ -106,7 +112,7 @@ def test_get_task(headers):
 def test_get_task_not_exists(headers):
 	hash_id = "IDONTEXIST"
 	url = "{}/{}".format(TASKS_URL, hash_id)
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	print(resp.content)
 	assert resp.status_code == 404
 
@@ -122,7 +128,7 @@ def test_update_task_formdata(headers, status, msg, expected_response_code):
 	url = "{}/{}".format(TASKS_URL, hash_id)
 
 	#FORM data
-	resp = requests.put(url, headers=headers, data={'status': status, 'msg': msg}, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.put(url, headers=headers, data={'status': status, 'msg': msg}, verify=False)
 	assert resp.status_code == expected_response_code
 
 
@@ -138,7 +144,7 @@ def test_update_task_jsondata(headers, status, msg, expected_response_code):
 
 	#JSON data
 	json={"status": status, "msg": msg}
-	resp = requests.put(url, headers=headers, json=json, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.put(url, headers=headers, json=json, verify=False)
 	assert resp.status_code == expected_response_code
 
 
@@ -148,7 +154,7 @@ def test_delete_task_id_exists(headers):
 	resp = create_task(headers)
 	hash_id = resp.json()["hash_id"]
 	url = "{}/{}".format(TASKS_URL, hash_id)
-	resp = requests.delete(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.delete(url, headers=headers, verify=False)
 	assert resp.status_code == 204
 
 
@@ -157,7 +163,7 @@ def test_delete_task_id_exists(headers):
 def test_delete_task_id_not_exists(headers):
 	hash_id = "IDONTEXIST"
 	url = "{}/{}".format(TASKS_URL, hash_id)
-	resp = requests.delete(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.delete(url, headers=headers, verify=False)
 	assert resp.status_code == 404 and "error" in resp.json()
 
 
@@ -167,7 +173,7 @@ def test_expire_task(headers):
 	resp = create_task(headers)
 	hash_id = resp.json()["hash_id"]
 	url = "{}/expire/{}".format(TASKS_URL, hash_id)
-	resp = requests.post(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.post(url, headers=headers, verify=False)
 	assert resp.status_code == 200 and "success" in resp.json()
 
 
@@ -176,14 +182,14 @@ def test_expire_task(headers):
 def test_expire_task_id_not_exists(headers):
 	hash_id = "IDONTEXIST"
 	url = "{}/expire/{}".format(TASKS_URL, hash_id)
-	resp = requests.post(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.post(url, headers=headers, verify=False)
 	assert resp.status_code == 404 and "error" in resp.json()
 
 
 @skipif_uses_gateway
 def test_status(headers):
 	url = f"{TASKS_URL}/status"
-	resp = requests.get(url, headers=headers, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, headers=headers, verify=False)
 	assert resp.status_code == 200
 
 
@@ -191,14 +197,14 @@ def test_status(headers):
 def test_taskslist():
 	url = f"{TASKS_URL}/taskslist"
 	json = {"service": "storage", "status_code":[]}
-	resp = requests.get(url, json=json, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, json=json, verify=False)
 	assert resp.status_code == 200
 
 @skipif_not_uses_gateway
 def test_taskslist():
 	url = f"{TASKS_URL}/taskslist"
 	json = {"service": "storage", "status_code":[]}
-	resp = requests.get(url, json=json, verify= (f"{SSL_PATH}{SSL_CRT}" if USE_SSL else False))
+	resp = requests.get(url, json=json, verify=False)
 	assert resp.status_code == 401